# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoCRM API",
    "description": "Autotherm project lifecycle system. Money is integer minor units with an explicit currency; quantities are decimal strings. PATCH: omit a field to keep it, send null to clear it.",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "description": "Stage key.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "open",
            "in": "query",
            "description": "Only leads not in a terminal stage.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_LeadSummary"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lead"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/leads/{id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lead"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/leads/{id}/stage": {
      "post": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_change_stage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageChange"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/leads/{id}/convert": {
      "post": {
        "tags": [
          "leads"
        ],
        "operationId": "leads_convert",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "`partner_id` may be omitted when the lead has one; `title` defaults to the lead's.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Currency": {
        "type": "string",
        "enum": [
          "HUF",
          "EUR"
        ]
      },
      "CurrentStage": {
        "type": "object",
        "required": [
          "stage_key",
          "entered_at"
        ],
        "properties": {
          "stage_key": {
            "type": "string"
          },
          "entered_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `validation`, `stage_gate`, `currency_locked`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. For `validation` it names the field and the reason."
          }
        }
      },
...
      "ItemBody": {
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unit_price"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "string",
            "description": "Decimal string, e.g. `\"2.5\"`."
          },
          "unit_price": {
            "type": "integer",
            "format": "int64",
            "description": "Minor units (fillér / eurocent). May be negative for discount lines."
          }
        }
      },
...
      "Items_LeadSummary": {
        "type": "object",
        "description": "The list envelope: `{\"items\": [...]}`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "title",
                "stage_key",
                "stage_label",
                "stage_entered_at",
                "created_at"
              ],
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "title": {
                  "type": "string"
                },
                "partner_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "partner_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "contact_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "contact_email": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "source": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "assigned_to": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "assigned_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "stage_key": {
                  "type": "string"
                },
                "stage_label": {
                  "type": "string"
                },
                "stage_entered_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "order_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "order_number": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "Lead": {
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "partner_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "contact_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "contact_phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "assigned_to": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "created_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "minicrm_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LeadBody": {
        "type": "object",
        "description": "Create requires `title`. On PATCH every field is optional; `null` clears.",
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "partner_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "contact_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "contact_phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "assigned_to": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "LeadDetail": {
        "type": "object",
        "required": [
          "lead",
          "history"
        ],
        "properties": {
          "lead": {
            "$ref": "#/components/schemas/Lead"
          },
          "stage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CurrentStage"
              }
            ]
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageEntry"
            }
          },
          "order": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrderRef"
              }
            ]
          }
        }
      },
...
      "Order": {
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "partner_id",
          "currency",
          "valuation_date",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "partner_id": {
            "type": "integer",
            "format": "int64"
          },
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "lead_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "project_type_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "valuation_date": {
            "type": "string",
            "format": "date"
          },
          "vehicle_make": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_plate": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_vin": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "assigned_to": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "created_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "minicrm_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderBody": {
        "type": "object",
        "required": [
          "currency"
        ],
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "partner_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "project_type_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "valuation_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date",
            "description": "Defaults to today. Reports normalise the order's value at this date's MNB rate."
          },
          "vehicle_make": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_plate": {
            "type": [
              "string",
              "null"
            ]
          },
          "vehicle_vin": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "assigned_to": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemBody"
            }
          }
        }
      },
...
      "OrderRef": {
        "type": "object",
        "required": [
          "id",
          "number"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          }
        }
      },
...
      "StageBody": {
        "type": "object",
        "required": [
          "stage"
        ],
        "properties": {
          "stage": {
            "type": "string",
            "description": "Target stage key."
          },
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required for backward moves and reopening a terminal stage."
          }
        }
      },
      "StageChange": {
        "type": "object",
        "required": [
          "from",
          "to",
          "kind"
        ],
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TransitionKind"
          }
        }
      },
...
      "StageEntry": {
        "type": "object",
        "required": [
          "id",
          "stage_key",
          "label_hu",
          "entered_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "stage_key": {
            "type": "string"
          },
          "label_hu": {
            "type": "string"
          },
          "entered_at": {
            "type": "string",
            "format": "date-time"
          },
          "left_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "entered_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "entered_by_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TransitionKind": {
        "type": "string",
        "enum": [
          "forward",
          "backward",
          "exit",
          "reopen"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "autocrm_session"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "bearer": []
    }
  ],
...
}